+++ src/lib/LanguageFlash.svelte
@@ -8,6 +8,11 @@
   
   export let vocabData: Word[] = [{pinyin:'...loading', traditional:'...loading', simplified:'...loading', meaning:'...loading'}]
 
+  const GREEN = '#4CAF50'
+  const LIGHT_GREEN = '#99ee99'
+  const RED = '#FF5252'
+  const LIGHT_RED = '#ee9999'
+
   let threshold = 2
   let numberWords = 3
   let mode = 'pinyin'
@@ -15,6 +20,8 @@
   let round = 0
   let mastered = false
   let reveal = false
+  let correct = false
+  let incorrect = false
   let cache = [...Array(vocabData.length)].map(x => 0)
   let unlearnedWords = vocabData.slice(0, numberWords)
   let leftToLearn = numberWords
@@ -39,10 +46,16 @@
   }
   const answer = (answer: boolean) => {
     if (answer) {
+      reveal = false
+      correct = true
       cache[randomNum] += 1
       if (cache[randomNum] === threshold) mastered = true
     }
-    else cache[randomNum] = Math.max(0, cache[randomNum] - 1)
+    else {
+      reveal = false
+      incorrect = true
+      cache[randomNum] = Math.max(0, cache[randomNum] - 1)
+    }
     cache = cache
     setTimeout(() => {
       const currentWords = unlearnedWords.map(x => vocabData.indexOf(x))
@@ -63,13 +76,20 @@
         leftToLearn = numberWords
       }
       mastered = false
-      reveal = false
+      correct = false
+      incorrect = false
     }, 300)
     // console.log(cache, randomNum, unlearnedIndex)
   }
-</script>
+  </script>
 
-<main>
+  <head>
+  <link
+    rel="stylesheet"
+    href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css"
+  />
+  </head>
+  <main>
   <div id="info">
     <div style="display: flex; justify-content: space-between; width: 100%; align-items: baseline;">
       <h3>ROUND: {round + 1}</h3>
@@ -110,28 +130,95 @@
   </div>
 
   <div id="word">
-    <h1 class={mastered ? 'mastered' : ''}>{unlearnedWords[unlearnedIndex][mode]}</h1>
-    <h2 style="line-height: 36px;">SCORE: {#if mastered} &#128170; {:else} {cache[randomNum]} {/if}</h2>
+    <h1 style="grid-column-start: 2; margin-bottom: 0px;" class={mastered ? 'mastered' : ''}>{unlearnedWords[unlearnedIndex][mode]}</h1>
+    {#if mastered} 
+      <h1 style="margin: 0px;" class="animate__animated animate__fadeOutUp">&#128170;</h1>
+    {/if} 
+    <h2 id="current-score">
+      SCORE: <span id="score-number">{cache[randomNum]}</span>
+    </h2>
+    <span id="point-marker">
+      {#if correct}
+        <h4 class="animate__animated animate__fadeOutUp" style={`color: ${GREEN}; margin: 0px;`}>+1</h4>
+      {:else if incorrect}
+        <h4 class="animate__animated animate__fadeOutUp" style={`color: ${RED}; margin: 0px;`}>-1</h4>
+      {/if}
+    </span>
   </div>
 
-  <div id="answer" style="min-height: 200px; display: flex; flex-direction: column; justify-content: space-evenly;">
+  <div
+    id="answer"
+    style="min-height: 200px; display: flex; flex-direction: column; justify-content: space-evenly;"
+  >
     {#if reveal}
-      <h3 style="margin: 0px 20px; align-self: center;">{unlearnedWords[unlearnedIndex].meaning}</h3>
-      <button style="border-color: #4CAF50; background-color: #99ee99; color: #000; align-self: center;" on:click={() => answer(true)}>CORRECT</button>
-      <button style="border-color: #FF5252; background-color: #ee9999; color: #000; align-self: center;" on:click={() => answer(false)}>INCORRECT</button>
-    {:else}
-      <button on:click={() => reveal = true} style="align-self: center;">
+      <h3
+        class="animate__animated animate__fadeIn"
+        style="margin: 0px 20px; align-self: center;"
+      >
+        {unlearnedWords[unlearnedIndex].meaning}
+      </h3>
+      <button
+        class="animate__animated animate__fadeInUp" 
+        style={`
+          border-color: ${GREEN};
+          background-color: ${LIGHT_GREEN};
+          color: #000;
+          align-self: center;
+        `}
+        on:click={() => answer(true)}
+      >
+        CORRECT
+      </button>
+      <button
+        class="animate__animated animate__fadeInUp"
+        style={`
+          border-color: ${RED};
+          background-color: ${LIGHT_RED};
+          color: #000;
+          align-self: center;
+        `}
+        on:click={() => answer(false)}
+      >
+        INCORRECT
+      </button>
+    {:else if !incorrect && !correct}
+      <button
+        class="animate__animated animate__fadeInUp"
+        on:click={() => reveal = true} style="align-self: center;"
+      >
         DEFINITION
       </button>
     {/if}
   </div>
-</main>
+  </main>
 
-<style>
+  <style>
 
   :root {
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
       Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  }
+
+  #word {
+    display: grid;
+    line-height: 36px;
+    grid-template-columns: repeat(3, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+    margin: auto;
+    width: fit-content;
+  }
+
+  #current-score {
+    grid-row-start: 2;
+    grid-column-start: 2;
+    grid-column-end: 3;
+  }
+
+  #point-marker {
+    grid-row-start: 2;
+    grid-column-start: 3;
+    grid-column-end: 4;
+    text-align: left;
   }
 
   #info {