+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [v1.0.6] - 2022-11-28
+
+### Added
+
+- add animation with animate.css
+- update score styling to use css grid
+- show +1/-1 point marker
+
 ## [v1.0.2] - 2022-11-22
 
 ### Added